+++ refactor/dispatcher/app/core/config.py
@@ -20,7 +20,10 @@
 
 from typing import List, Union
 
+from dotenv import find_dotenv, load_dotenv
 from pydantic import AnyHttpUrl, BaseSettings, validator
+
+load_dotenv()
 
 
 class Settings(BaseSettings):
@@ -45,6 +48,10 @@ class Settings(BaseSettings):
 
     class Config:
         case_sensitive = True
+        env_file = ".env"
 
 
 settings = Settings()
+
+print("Configuration:")
+print(settings)