+++ refactor/ui_backend/app/core/config.py
@@ -21,14 +21,16 @@
 import os
 from typing import List, Union
 
+from dotenv import find_dotenv, load_dotenv
 from pydantic import AnyHttpUrl, BaseSettings, validator
+
+load_dotenv()
 
 
 class Settings(BaseSettings):
     API_V0_STR: str = "/api/v0"
     BACKEND_CORS_ORIGINS: List[AnyHttpUrl] = []
-    DATA_OS_SVC_HOST_URI: str = os.getenv("DATA_OS_SVC_HOST_URI")
-    UI_SVC_PORT: int = 8004
+    UI_SVC_PORT: int = 8005
     UI_SVC_HOST: str = "localhost"
 
     @validator("BACKEND_CORS_ORIGINS", pre=True)
@@ -41,6 +43,10 @@ class Settings(BaseSettings):
 
     class Config:
         case_sensitive = True
+        env_file = ".env"
 
 
 settings = Settings()
+
+print("Configuration:")
+print(settings)